[Extension] Rename extension file according to Angular conventions
Although the extension files does not use angular framework, we rename
the files to match angular file naming conventions for consistent
styling across the repo.
