+++ diagnostics-extension/src/services/fake-diagnostics.service.ts
