+++ diagnostics-extension/src/services/fake-events.service.ts
