+++ diagnostics-extension/src/services/fake-telemetry.data.ts
